requirements.txt
python-telegram-bot==13.7
pytest==6.2.5
requests==2.26.0
google-cloud-storage==1.42.3